:recycle: Move to sort.go for sort functions
